Use size_t for indexes and collection capacity

+++ src/collections.h
@@ -20,7 +20,7 @@ typedef enum
 // == COMMON ==================================================================
 
 // Return the number of items in the array
-int clxns_count(void *collection);
+size_t clxns_count(void *collection);
 
 // Create a new iterator
 void *clxns_iter_new(void *collection);
@@ -37,19 +37,19 @@ void clxns_iter_free(void *iterator);
 // == RESIZE ARRAY =============================================================
 
 // Create and return a new array. Specify the initial size.
-void *resize_array(int init_size);
+void *resize_array(size_t init_size);
 
 // Add an item to the array
 void resize_array_add(void *array, void *item);
 
 // Access an item at the given position in the array
-C_STATUS resize_array_get(void *array, int index, void **item);
+C_STATUS resize_array_get(void *array, size_t index, void **item);
 
 // Swap two items in the array
-C_STATUS resize_array_exchange(void *array, int first, int second);
+C_STATUS resize_array_exchange(void *array, size_t first, size_t second);
 
 // Remove an item from the array, set item to the value if non-zero is passed in
-C_STATUS resize_array_remove(void *array, int index, void **item);
+C_STATUS resize_array_remove(void *array, size_t index, void **item);
 
 // Shallow copy a resize array
 void *resize_array_copy(void *array);
@@ -64,8 +64,8 @@ void resize_array_free(void *array, int items);
  * size and a callback to compare items
  * _min orders ascending, _max orders descending
  */
-void *priority_queue_min(int init_size, int (*compare)(void *first, void *second));
-void *priority_queue_max(int init_size, int (*compare)(void *first, void *second));
+void *priority_queue_min(size_t init_size, int (*compare)(void *first, void *second));
+void *priority_queue_max(size_t init_size, int (*compare)(void *first, void *second));
 
 // Add an item to the priority queue. May return CE_NULL_ITEM.
 C_STATUS priority_queue_add(void *pqueue, void *item);
@@ -85,7 +85,7 @@ void priority_queue_free(void *array, int items);
 // == HASH TABLE ==============================================================
 
 // Create and return a new hash table. Specify the initial size.
-void *hash_table(int init_size);
+void *hash_table(size_t init_size);
 
 // Associate a key with a value
 void hash_table_add(void *table, char *key, void *value);
